fastmcp_docs_server/requirements.txt
fastmcp
PyYAML
supabase
uvicorn
starlette
python-dotenv
